Spring Boot with all functionalties

+++ labs/springboot/sboot-rest-employee/src/main/java/com/sboot/empapp/service/EmployeeServiceImpl.java
@@ -1,7 +1,10 @@
 package com.sboot.empapp.service;
 
 import java.util.List;
+import java.util.Optional;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,32 +14,53 @@ import com.sboot.empapp.model.Employee;
 
 @Service
 public class EmployeeServiceImpl implements EmployeeService {
+	
+	Logger logger = LoggerFactory.getLogger(EmployeeServiceImpl.class);
 	@Autowired
 	EmployeeRepositry repo;
 
+	@Override
 	public Employee create(Employee employee) throws EmployeeException {
-
 		return repo.save(employee);
+
 	}
 
+	@Override
 	public List<Employee> list() throws EmployeeException {
-           System.out.println("I am also in list method");
+
 		return repo.findAll();
 	}
 
-	public Employee get(int id) {
-
-		return repo.findById(id).get();
+	@Override
+	public Employee get(int id) throws EmployeeException {
+		Optional<Employee> optional = repo.findById(id);
+		Employee employee = null;
+		if (optional.isPresent()) {
+			employee = optional.get();
+		} else {
+			logger.error("Employee id not found for the given id " + id);
+			throw new EmployeeException(" Employee not found for id :: " + id);
+		}
+		return employee;
 	}
 
-	public Employee update(int id, Employee employee) {
-
-		return repo.save(employee);
+	@Override
+	public Employee update(int id, Employee employee) throws EmployeeException {
+		if (get(id) != null)
+			return repo.save(employee);
+		else
+			logger.error("Employee id not found for the given id " + id);
+		throw new EmployeeException(" Employee not found for id :: " + id);
 	}
 
-	public void delete(int id) {
-
-		repo.deleteById(id);
+	@Override
+	public void delete(int id) throws EmployeeException {
+		if (get(id) != null)
+			repo.deleteById(id);
+//		else
+//			logger.error("Employee id not found for the given id " + id);
+//			throw new EmployeeException(" Employee not found for id :: " + id);
+//			
 	}
 
 }
